requirements.txt
requests==2.5.1
GitPython==2.0.7
